fixed bug with locking geo at 0.0\0.0 changed types of similar values

+++ app/src/main/java/app/seals/weather/network/pojo/HourlyPOJO.kt
@@ -10,5 +10,5 @@ data class HourlyPOJO (
     @SerializedName("weathercode"          ) var weathercode         : ArrayList<Int>    = arrayListOf(0),
     @SerializedName("surface_pressure"     ) var surfacePressure     : ArrayList<Double> = arrayListOf(0.0),
     @SerializedName("windspeed_10m"        ) var windspeed10m        : ArrayList<Double> = arrayListOf(0.0),
-    @SerializedName("winddirection_10m"    ) var winddirection10m    : ArrayList<Int>    = arrayListOf(0)
+    @SerializedName("winddirection_10m"    ) var winddirection10m    : ArrayList<Double>    = arrayListOf(0.0)
 )
